New Abstract Class
This abstract class is the parent for all the services. It implements
new interface from TestProjectCommon - ICommonService

Some common methods that can be used in all the services were moved to
this class:
1) getQuery (protected)
2) getQuestionWithDelimeters (protected)
3) getAnswers (protected)
4) getQuestionById (public)
5) encodeImage (protected)
Some new methods were added:
1) getLimitsForQuery  and getLimitsForQueryWithWhere - these methods add
restrictions to the query (CompanyId) to the query. The first  is
abstract here.
2) getAllMetaCategoriesFromDataBase(), getAllCategories1FromDataBase(),
getAllCategories2FromDataBase() - return lists of done categories with
restrictions according to the user's access (Maintenance, Company or
User)

3) ifAllowed (abstract) - checks if the user (or Company) is allowed to
get this question-item (EnitityQuesitionAttributes)